1.13.3
У комментариев и добавляемых страниц сайта появилась возможность
добавлять визуальный редактор со вставкой фото, ссылок, видео и файлов.
Настраивается в настройке раздела.
Улучшена безопасность.
К добавляемым страницам добавлена капча.

+++ ed2/image_upload.php
@@ -1,4 +1,7 @@
 <?php
+// Настройка
+$folder =  '../img/'; //директория в которую будет загружен файл
+
 if(!eregi('image/', $_FILES['file']['type'])) exit(0);
 
 foreach ($_FILES['file'] as $secvalue) {
@@ -24,34 +27,54 @@ foreach ($_FILES['file'] as $secvalue) {
     }
 }
 
-$folder =  '../img/';//директория в которую будет загружен файл
-$type = str_replace("image/","",strtolower($_FILES['file']['type']));
-if ($type == 'png' || $type == 'jpg' || $type == 'gif' || $type == 'jpeg' || $type == 'pjpeg') {
-  $foto =  md5(date('YmdHis')).'.'.$type;
-  if (is_uploaded_file($_FILES['file']['tmp_name'])) {
-    if (move_uploaded_file($_FILES['file']['tmp_name'], $folder.$foto)) {
-      if (extension_loaded('imagick') && class_exists("Imagick")) {
-        $image = new Imagick($folder.$foto);
-        // сжатие
-        list($width, $height, $type, $attr) = getimagesize($folder.$foto);
-        if ($width > 1000) $image->thumbnailImage(1000,0); // по горизонтали до 1000 пикселей
-        else if ($height > 1200) $image->thumbnailImage(0,1200); // по вертикали до 1200 пикселей
-        // наводим резкость, если превью мелкое
-        if ($width < 300) $image->sharpenImage(4, 1);
-        //закругляем углы
-        $thumb->roundCorners(5, 5);
-        // ориентация фото
-        $orientation = exif_read_data($folder.$foto);
-        if ($orientation['Orientation'] !== 0 && $orientation['Orientation'] !== 1 && $orientation['Orientation'] != "") {
-            $degres = ($orientation['Orientation']- 1) * 90; 
-            $image->rotateImage('', $degres);
+if (!empty($_FILES['file']['name'])) {
+  $type = str_replace("image/","",strtolower($_FILES['file']['type']));
+  if ($type == 'png' || $type == 'jpg' || $type == 'gif' || $type == 'jpeg' || $type == 'pjpeg') {
+    $foto =  md5(date('YmdHis')).'.'.$type;
+    if (is_uploaded_file($_FILES['file']['tmp_name'])) {
+      if (move_uploaded_file($_FILES['file']['tmp_name'], $folder.$foto)) {
+        // Обработка фото (при наличии библиотеки Imagick)
+        if (extension_loaded('imagick') && class_exists("Imagick")) {
+          $image = new Imagick($folder.$foto);
+          // сжатие
+          list($width, $height, $type, $attr) = getimagesize($folder.$foto);
+          if ($width > 1000) $image->thumbnailImage(1000,0); // по горизонтали до 1000 пикселей
+          else if ($height > 1200) $image->thumbnailImage(0,1200); // по вертикали до 1200 пикселей
+          // наводим резкость, если превью мелкое
+          if ($width < 300) $image->sharpenImage(4, 1);
+          //закругляем углы
+          $thumb->roundCorners(5, 5);
+          // ориентация фото
+          $orientation = exif_read_data($folder.$foto);
+          if ($orientation['Orientation'] !== 0 && $orientation['Orientation'] !== 1 && $orientation['Orientation'] != "") {
+              $degres = ($orientation['Orientation']- 1) * 90; 
+              $image->rotateImage('', $degres);
+          }
+          $image->writeImage();
+          $image->destroy();
         }
-        $image->writeImage();
-        $image->destroy();
       }
+      $array = array('filelink' => str_replace("..","",$folder).$foto);
+      if (is_image($folder.$foto)) 
+        echo stripslashes(json_encode($array)); 
     }
-    $array = array('filelink' => '/img/'.$foto);
-    echo stripslashes(json_encode($array)); 
   }
+}
+
+// проверка изображения на безопасность
+function is_image($image_path) {
+    if (!$f = fopen($image_path, 'rb')) {
+        return false;
+    }
+    $data = fread($f, 8);
+    fclose($f);
+    // signature checking
+    $unpacked = unpack("H12", $data);
+    if (array_pop($unpacked) == '474946383961' || array_pop($unpacked) == '474946383761') return "gif";
+    $unpacked = unpack("H4", $data);
+    if (array_pop($unpacked) == 'ffd8') return "jpg";
+    $unpacked = unpack("H16", $data);
+    if (array_pop($unpacked) == '89504e470d0a1a0a') return "png";
+    return false;
 }
 ?>
